+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added support for Solana.
 - Added new loading animation for pending orders in navbar capsule
+- Added dynamic price impact in the fee tooltip.
 
 ### Fixed
 
@@ -22,6 +23,8 @@
 - Fixed Notification bubble Wrap
 - Fixed Modal dropdown overflow for stake duration.
 - Fixed Network Fee for Bitcoin.
+- Fixed Total cost in Fee tooltip.
+- Fixed the Price impact logic in fee tooltip.
 
 ## [1.0.2]